Remove unnecessary outdated files. (#200)
* Remove unnecessary outdated files for simple gradio demo and safe trainer saver.

* Remove incomplete sentences in the readme.

* Remove the finetuning script that uses the HF trainer, since it is not being used.
